Sample Differential MPC objects and I/O Midpoints
The following .mat files are sample differential MPC objects and midpoint values for the values going into the Multiple MPC block and the values coming out. There are files for an MPC for the throttle and files for the wastegate also. In the RunMe to create the throttle MPCs, the following parameters were used:
   ThrB = [3 3 6 9];        % Throttle Base value
   ThrW = [1.2 1.5 1.8 2];  % Throttle wiggle value
   WgB = [100 100 100 100]; % Wastegate Base value (held constant at 100% open)
   WgW = [0 0 0 0];         % Wastegate wiggle value
   rpm = [2000 3000 4000 5000]; % engine speed values (operating points)

In the RunMe to create the wastegate MPCs, the following parameters were used. 
   ThrB = [100 100 100 100];     % Throttle Base value
   ThrW = [0 0 0 0];      % Throttle wiggle value
   WgB = [62 56 57 57];   % Wastegate Base value (held constant at 100% open)
   WgW = [10 9 9 8];       % Wastegate wiggle value
   rpm = [2000 3000 4000 5000]; % engine speed values (operating points)

This zip folder contains the following .mat files:
   MidActuator_Thr
   MidMAP_Thr
   MPCobjs_Thr
   MidActuator_Wg
   MidMAP_Wg
   MPCobjs_Wg

See the OperationsManual for their implementation. A list of more throttle and wiggle values to produce good throttle-MAP models can be found in the Final System Review April 2021.